Edited _Layout.cshtml file to add a UI to the home page.
Added _ModelIndexLayout.cshtml to keep the layout consistent between the movie, actor, edit and create pages.

Added some scripts to adjust the layout according to window size.

The three bottom links on the homepage don't work yet. And I want to add a splash screen so the startup isn't so choppy.